logout

+++ frontend/src/components/Layout.js
@@ -43,7 +43,7 @@ function Layout({children}) {
             icon: 'ri-home-4-line'
         },
         {
-            name: 'Dashboard',
+            name: 'Admin Dashboard',
             path: '/users',
             icon: 'ri-user-settings-line'
         },
@@ -117,17 +117,17 @@ function Layout({children}) {
         },
         {
             name: 'Stock',
-            path: '/dashboard/stock',
+            path: '/dashboard/logistics/stock',
             icon: 'ri-archive-line',
         },
         {
             name: 'Suppliers',
-            path: '/supplier',
+            path: '/dashboard/logistics/supplier',
             icon: 'ri-truck-line', 
         },
         {
             name: 'Purchase Orders',
-            path: '/purchaseOrder',
+            path: '/dashboard/logistics/purchaseOrder',
             icon: 'ri-store-2-line', 
         }
     ];
@@ -135,14 +135,14 @@ function Layout({children}) {
     //5. staff manager
     const staffManagerMenu = [
         {
-            name: 'Home',
-            path: '/',
-            icon: 'ri-home-4-line'
-        },
-        {
             name: 'Dashboard',
             path: '/employee',
             icon: 'ri-user-settings-line'
+        },
+        {
+            name: 'Add Employee',
+            path: '/signup',
+            icon: 'ri-add-circle-line'
         },
         {
             name: 'Attendance',
@@ -168,7 +168,6 @@ function Layout({children}) {
             path: '/dashboard/trainee',
             icon: 'ri-home-4-line'
         },
-        
         {
             name: 'Employee Management',
             path: '/dashboard/employee',
@@ -214,10 +213,13 @@ function Layout({children}) {
     }
 
     const handleLogout = () => {
+        const confirmLogout = window.confirm('Are you sure you want to logout?');
         // Show a confirmation message before logging out
-        localStorage.clear();
-        dispatch(setUser(null)); // Clear user state in Redux store
-        navigate('/login');
+        if (confirmLogout) {
+            localStorage.clear();
+            dispatch(setUser(null)); // Clear user state in Redux store
+            navigate('/login');
+        }
     };
 
     const getData = async () => {
@@ -232,7 +234,6 @@ function Layout({children}) {
             console.log(error);
         }
     };
-
 
     // Conditionally render children based on user role
     const renderChildren = () => {
